requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn==1.6.1==1.6.1
scapy
tensorflow
torch
streamlit
imbalanced-learn
jupyter
joblib
pyyaml
paho-mqtt
fastapi
uvicorn[standard]
plotly
websockets
twilio
python-dotenv
psutil
tqdm
xgboost
lightgbm
optuna
sqlalchemy>=2.0.0